@@ -35,9 +35,9 @@
         }
     ],
     "require": {
-        "php": "^7.4|^8.0|^8.1|^8.2",
-        "guzzlehttp/guzzle": "^7.2.0",
-        "ext-json": "*"
+        "ext-json": "*",
+        "php": "^8.1|^8.2",
+        "guzzlehttp/guzzle": "^7.2.0"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^2.16",
