@@ -1,13 +1,17 @@
 <?php
 
+declare(strict_types=1);
+
 namespace NotchPay\Exceptions;
 
-class ApiException extends NotchPayException
+final class ApiException extends NotchPayException
 {
-    public $errors;
+    public array $errors = [];
+
     public function __construct($message, array $errors = [])
     {
         parent::__construct($message);
+
         $this->errors = $errors;
     }
 }
