@@ -1,19 +1,27 @@
 <?php
 
+declare(strict_types=1);
+
 namespace NotchPay;
 
 use NotchPay\Exceptions\InvalidArgumentException;
 
-class NotchPay
+final class NotchPay
 {
-    /** @var string The Notch Pay API key to be used for requests. */
+    /**
+     * The Notch Pay API key to be used for requests.
+     */
     public static string $apiKey;
 
-    /** @var string The instance API key, settable once per new instance */
-    private $instanceApiKey;
+    /**
+     * The instance API key, settable once per new instance
+     */
+    private string $instanceApiKey;
 
-    /** @var string The base URL for the Notch Pay API. */
-    public static $apiBase = 'https://api.notchpay.co';
+    /**
+     * The base URL for the Notch Pay API.
+     */
+    public static string $apiBase = 'https://api.notchpay.co';
 
     /**
      * @return string the API key used for requests
@@ -26,22 +34,36 @@ class NotchPay
     /**
      * Sets the API key to be used for requests.
      *
+     * @throws InvalidArgumentException
      */
     public static function setApiKey(string $apiKey): void
     {
         self::validateApiKey($apiKey);
+
         self::$apiKey = $apiKey;
     }
 
-    private static function validateApiKey($apiKey): bool
+    /**
+     * @param string $apiKey
+     * @return bool
+     *
+     * @throws InvalidArgumentException
+     */
+    private static function validateApiKey(string $apiKey): bool
     {
-        if ($apiKey == '' || ! is_string($apiKey)) {
+        if (empty($apiKey)) {
             throw new InvalidArgumentException('Api key must be a string and cannot be empty');
         }
 
-        if(substr( $apiKey, 0, 2 ) !== "b." && substr( $apiKey, 0, 3 ) !== "sb."  && substr( $apiKey, 0, 3 ) !== "pk." && substr( $apiKey, 0, 8 ) !== "pk_test.") {
+        if (
+            !str_starts_with($apiKey, 'b.')
+            && !str_starts_with($apiKey, 'sb.')
+            && !str_starts_with($apiKey, 'pk.')
+            && !str_starts_with($apiKey, 'pk_test.')
+        ) {
             throw new InvalidArgumentException('Api key must have a valid signature.');
         }
+
         return true;
     }
 }
