@@ -1,7 +1,9 @@
 <?php
 
+declare(strict_types=1);
+
 namespace NotchPay\Exceptions;
 
-class InvalidArgumentException extends \Exception
+final class InvalidArgumentException extends \Exception
 {
 }
